Separate step filter from pipeline generation
